bug: use uint32_t for IP numbers
The commit 1e82461875373748d619acd72791d0af2d2cb621 was
incomplete, and at the end was a cause of off by one miscount.
The reason was bsearch where long int needed to be unsigned.
Weird that the original had work while there where so many type
mismatches.

+++ src/analyze.c
@@ -64,7 +64,7 @@ int prepare_data(void)
 	for (i = 0; i < num_touches; i++) {
 		if (bsearch
 		    (&touches[i], leases, (size_t) num_leases,
-		     sizeof(long int), &intcomp) == NULL) {
+		     sizeof(uint32_t), &intcomp) == NULL) {
 			touches[j] = touches[i];
 			j++;
 		}
